+++ src/ports/web/api/count.rs
@@ -2,13 +2,10 @@ use super::*;
 
 #[get("/count/entries")]
 pub fn get_count_entries(db: DbConn) -> Result<usize> {
-    // TODO: Replace with count query
-    let entries = db.read_only()?.all_entries()?;
-    Ok(Json(entries.len()))
+    Ok(Json(db.read_only()?.count_entries()?))
 }
 
 #[get("/count/tags")]
 pub fn get_count_tags(db: DbConn) -> Result<usize> {
-    // TODO: Replace with count query
-    Ok(Json(db.read_only()?.all_tags()?.len()))
+    Ok(Json(db.read_only()?.count_tags()?))
 }